Fix Held Keys bitmasks through native host feeders
soemdsp_graph_set_param now takes double so 88-key Held Keys masks survive Bias host CV feeders (float32 only keeps exact ints through 2^24; middle-C is bit 36). Unblocks Arp over Keyboard Held Keys.

+++ native_modules/graph_engine/graph_engine.cpp
@@ -9195,7 +9195,9 @@ extern "C" int soemdsp_graph_snap_controls(int handle) {
   return 0;
 }
 
-extern "C" int soemdsp_graph_set_param(int handle, unsigned int nodeHash, int paramId, float value) {
+// double — Held Keys / Scale bitmasks must survive host→Bias feeders (float32
+// only keeps exact ints through 2^24; middle-C is bit 36).
+extern "C" int soemdsp_graph_set_param(int handle, unsigned int nodeHash, int paramId, double value) {
   Circuit* g = get(handle);
   if (!g) return -1;
   const int idx = find_node(*g, nodeHash);
@@ -9203,7 +9205,7 @@ extern "C" int soemdsp_graph_set_param(int handle, unsigned int nodeHash, int pa
   if (!(value == value)) return 0;
   Control* c = control_for_param(g->nodes[idx], paramId);
   if (!c) return 0;
-  control_set_target(*g, *c, (double)value);
+  control_set_target(*g, *c, value);
   return 0;
 }
